pyproject.toml
[tool.poetry]
name = "llm-quality"
version = "0.1.0"
description = ""
authors = ["Evgenii Danilov <evgenii_danilov@epam.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.16"
pandas = "^2.2.2"
langchain-unstructured = "^0.1.2"
langchain-openai = "^0.1.23"
boto3 = "1.34.131"
python-dotenv = "^1.0.1"
jupyterlab = "^4.2.5"
fuzzywuzzy = "^0.18.0"
notebook = "^7.2.2"
openpyxl = "^3.1.5"
python-levenshtein = "^0.25.1"
deepeval = "^1.1.6"
langchain-aws = "^0.1.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
